fix (API) BUG-66: Reactivate policies

See merge request bugshot/web/api!461

+++ app/Http/Controllers/NotificationController.php
@@ -168,7 +168,7 @@ class NotificationController extends Controller
 	public function destroy(User $user, DatabaseNotification $notification, NotificationService $notificationService)
 	{
 		// Check if the user is authorized to delete the notification
-		// $this->authorize('delete', [DatabaseNotification::class, $notification]);
+		$this->authorize('delete', [DatabaseNotification::class, $notification]);
 
 		// Delete the notification
 		$val = $notificationService->delete($user, $notification);
@@ -240,7 +240,7 @@ class NotificationController extends Controller
 	public function destroyAll(User $user)
 	{
 		// Check if the user is authorized to delete the notifications
-		// $this->authorize('deleteAll', [DatabaseNotification::class, $user]);
+		$this->authorize('deleteAll', [DatabaseNotification::class, $user]);
 
 		// Delete the notifications
 		$val = $user->notifications()->delete();
